tests/ambient-light/w3c/AmbientLightSensor_secure_context.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>AmbientLightSensor Test: insecure context</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://www.w3.org/TR/ambient-light/">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div id="log"></div>
<h2>Precondition</h2>
<ol>
  <li>
    This is an insecurity context.
  </li>
</ol>
<script>

test(function() {
  assert_throws('SecurityError', function() { 
    let sensor = new AmbientLightSensor();
  });
}, "This test if throw SecurityError in an insecurity context.");

</script>
